tg-bot-revit-project/requirements.txt
altgraph==0.17.4
anyio==4.4.0
black==24.8.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
isort==5.13.2
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pefile==2023.2.7
platformdirs==4.2.2
pluggy==1.5.0
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
pytest==8.2.2
python-dotenv==1.0.1
python-telegram-bot==13.7
pywin32-ctypes==0.2.2
requests==2.32.3
rpws==1.1.0
six==1.16.0
sniffio==1.3.1
transliterate==1.10.2
urllib3==2.2.2
